chore: improve code quality
small adjustements and add javadoc
