+++ app/src/androidTest/java/com/android/unio/ui/notifications/NotificationTest.kt
@@ -31,99 +31,83 @@ import org.junit.Rule
 import org.junit.Test
 
 class NotificationTest : TearDown() {
-    @get:Rule
-    val composeTestRule = createComposeRule()
+  @get:Rule val composeTestRule = createComposeRule()
 
-    @get:Rule
-    val permissionRule = GrantPermissionRule.grant(android.Manifest.permission.POST_NOTIFICATIONS)
+  @get:Rule
+  val permissionRule = GrantPermissionRule.grant(android.Manifest.permission.POST_NOTIFICATIONS)
 
-    @MockK
-    private lateinit var navigationAction: NavigationAction
+  @MockK private lateinit var navigationAction: NavigationAction
 
-    private lateinit var userViewModel: UserViewModel
+  private lateinit var userViewModel: UserViewModel
 
-    // Mock event repository to provide test data.
-    @MockK
-    private lateinit var eventRepository: EventRepositoryFirestore
+  // Mock event repository to provide test data.
+  @MockK private lateinit var eventRepository: EventRepositoryFirestore
 
-    @MockK(relaxed = true)
-    private lateinit var searchRepository: SearchRepository
+  @MockK(relaxed = true) private lateinit var searchRepository: SearchRepository
 
-    @MockK
-    private lateinit var imageRepository: ImageRepositoryFirebaseStorage
+  @MockK private lateinit var imageRepository: ImageRepositoryFirebaseStorage
 
-    @MockK
-    private lateinit var userRepository: UserRepositoryFirestore
-    private lateinit var eventViewModel: EventViewModel
-    private lateinit var searchViewModel: SearchViewModel
-    private lateinit var context: Context
-    private val timerNotif: Long = 5000 // short time so test doesn't take too long
-    private val mockNotification = UnioNotification(
-        "my notification title", "super duper event, come it will be nice :)",
-        R.drawable.other_icon, "1234", "anonymous",0, System.currentTimeMillis() + timerNotif
-    )
+  @MockK private lateinit var userRepository: UserRepositoryFirestore
+  private lateinit var eventViewModel: EventViewModel
+  private lateinit var searchViewModel: SearchViewModel
+  private lateinit var context: Context
+  private val timerNotif: Long = 5000 // short time so test doesn't take too long
+  private val mockNotification =
+      UnioNotification(
+          "my notification title",
+          "super duper event, come it will be nice :)",
+          R.drawable.other_icon,
+          "1234",
+          "anonymous",
+          0,
+          System.currentTimeMillis() + timerNotif)
 
+  @Before
+  fun setUp() {
+    MockKAnnotations.init(this)
+    every { eventRepository.init(any()) } just runs
+    every { userRepository.init(any()) } just runs
+    context = InstrumentationRegistry.getInstrumentation().targetContext
+    searchViewModel = spyk(SearchViewModel(searchRepository))
+    eventViewModel = EventViewModel(eventRepository, imageRepository)
+    userViewModel = spyk(UserViewModel(userRepository))
+  }
 
-    @Before
-    fun setUp() {
-        MockKAnnotations.init(this)
-        every { eventRepository.init(any()) } just runs
-        every { userRepository.init(any()) } just runs
-        context = InstrumentationRegistry.getInstrumentation().targetContext
-        searchViewModel = spyk(SearchViewModel(searchRepository))
-        eventViewModel = EventViewModel(eventRepository, imageRepository)
-        userViewModel = spyk(UserViewModel(userRepository))
-
-
+  @Test
+  fun notificationIsSentTest() {
+    composeTestRule.setContent {
+      HomeScreen(navigationAction, eventViewModel, userViewModel, searchViewModel)
     }
+    NotificationWorker.schedule(context, mockNotification)
 
-    @Test
-    fun notificationIsSentTest() {
-        composeTestRule.setContent {
-            HomeScreen(navigationAction, eventViewModel, userViewModel, searchViewModel)
-        }
-        NotificationWorker.schedule(
-            context,
-            mockNotification
-        )
-
-        Thread.sleep(timerNotif + 500)
-        val manager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        with(manager.activeNotifications.first()) {
-            assertEquals(mockNotification.notificationId, this.id)
-            assertEquals(
-                mockNotification.title, this.notification.extras.getString("android.title")
-            )
-            assertEquals(
-                mockNotification.message,
-                this.notification.extras.getString("android.text")
-            )
-
-            assertEquals(mockNotification.icon, this.notification.smallIcon.resId)
-        }
-        manager.cancelAll()
+    Thread.sleep(timerNotif + 500)
+    val manager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+    with(manager.activeNotifications.first()) {
+      assertEquals(mockNotification.notificationId, this.id)
+      assertEquals(mockNotification.title, this.notification.extras.getString("android.title"))
+      assertEquals(mockNotification.message, this.notification.extras.getString("android.text"))
+      assertEquals(mockNotification.icon, this.notification.smallIcon.resId)
     }
+    manager.cancelAll()
+  }
 
-    @Test
-    fun notificationScheduledThenCanceled() {
-        composeTestRule.setContent {
-            HomeScreen(navigationAction, eventViewModel, userViewModel, searchViewModel)
-        }
-        NotificationWorker.schedule(
-            context,
-            mockNotification
-        )
-        val manager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        Thread.sleep(timerNotif / 2)
-        NotificationWorker.unschedule(context, mockNotification.notificationId)
-        Thread.sleep(timerNotif / 2 + 500)
-        assert(manager.activeNotifications.isEmpty())
+  @Test
+  fun notificationScheduledThenCanceled() {
+    composeTestRule.setContent {
+      HomeScreen(navigationAction, eventViewModel, userViewModel, searchViewModel)
     }
+    NotificationWorker.schedule(context, mockNotification)
+    val manager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+    Thread.sleep(timerNotif / 2)
+    NotificationWorker.unschedule(context, mockNotification.notificationId)
+    Thread.sleep(timerNotif / 2 + 500)
+    assert(manager.activeNotifications.isEmpty())
+  }
 
-    @After
-    override fun tearDown() {
-        super.tearDown()
-        val manager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        manager.cancelAll()
-    }
+  @After
+  override fun tearDown() {
+    super.tearDown()
+    val manager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+    manager.cancelAll()
+  }
 }
