+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
         android:theme="@style/Theme.SampleApp"
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
-        <receiver android:name=".model.notification.NotificationReceiver"
+        <receiver android:name=".model.notification.NotificationScheduler"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.BOOT_COMPLETED" />
